backend/requirement.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
pymupdf
pytesseract
python-docx
pillow
anthropic
openai
tenacity
streamlit
requests
python-dotenv
pytest